Fix cargo table SQL
